ARM: dts: msm8998: Disable unhandled or broken IRQ monitoring
This is primarily intended for generic x86 kernels which may run on all
sorts of broken systems. Our kernel will run on a very specific hardware
configuration with known drivers and devices, which makes this
unnecessary.

Enable noirqdebug to avoid the minor overhead added to every IRQ
handled.

Suggested-by: Tyler Nijmeh <tylernij@gmail.com>
Signed-off-by: kdrag0n <dragon@khronodragon.com>
(Adapted to msm8998)
Signed-off-by: crian <crian84@gmail.com>